@@ -1,12 +1,12 @@
 plugins {
-    id 'org.springframework.boot' version '2.5.5'
-    id 'io.spring.dependency-management' version '1.0.11.RELEASE'
+    id 'org.springframework.boot' version '3.2.0'
+    id 'io.spring.dependency-management' version '1.1.4'
     id 'java'
 }
 
 group = 'com.example'
 version = '0.0.1-SNAPSHOT'
-sourceCompatibility = '15'
+sourceCompatibility = '17'
 
 configurations {
     compileOnly {
@@ -31,16 +31,17 @@ dependencies {
     implementation 'org.springframework.boot:spring-boot-starter-web'
     implementation 'org.springframework.boot:spring-boot-starter-thymeleaf'
     implementation 'org.springframework.boot:spring-boot-starter-log4j2'
-    implementation group: 'org.apache.httpcomponents', name: 'httpclient', version: '4.5'
-    implementation 'org.apache.commons:commons-csv:1.9.0'
-    implementation group: 'commons-io', name: 'commons-io', version: '2.6'
-    implementation("com.google.guava:guava:31.0-jre")
+    implementation group: 'javax.annotation', name: 'javax.annotation-api', version: '1.3.2'
+    implementation group: 'org.apache.httpcomponents', name: 'httpclient', version: '4.5.13'
+    implementation 'org.apache.commons:commons-csv:1.10.0'
+    implementation group: 'commons-io', name: 'commons-io', version: '2.7'
+    implementation('com.google.guava:guava:31.1-jre')
     implementation group: 'org.apache.commons', name: 'commons-lang3', version: '3.0'
     compileOnly 'org.projectlombok:lombok'
     developmentOnly 'org.springframework.boot:spring-boot-devtools'
     annotationProcessor 'org.projectlombok:lombok'
     testImplementation 'org.springframework.boot:spring-boot-starter-test'
-    testImplementation "org.mockito:mockito-core:3.+"
+    testImplementation 'org.mockito:mockito-core:5.2.0'
 }
 
 test {
