@@ -8,7 +8,7 @@ import org.springframework.web.bind.annotation.RestController;
 import top.mobility.ip.service.IpCountryService;
 import top.mobility.ip.util.IpUtils;
 
-import javax.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletRequest;
 
 @RestController
 @RequestMapping("/ip-country/address/")
